#916 Renamed the package for the test-app classes to make it more distinctive for the purpose of moving it into a separate project at a later stage.
