--- platform-pojo-bl/src/main/java/fielden/close_leave/ITgCloseLeaveExampleDetail.java
@@ -1,13 +1,0 @@
-package fielden.close_leave;
-
-import ua.com.fielden.platform.dao.IEntityDao;
-
-/** 
- * Companion object for entity {@link TgCloseLeaveExampleDetail}.
- * 
- * @author Developers
- *
- */
-public interface ITgCloseLeaveExampleDetail extends IEntityDao<TgCloseLeaveExampleDetail> {
-
-}
