--- platform-pojo-bl/src/main/java/fielden/close_leave/TgCloseLeaveExampleDetail.java
@@ -1,28 +1,0 @@
-package fielden.close_leave;
-
-import fielden.close_leave.TgCloseLeaveExample;
-import ua.com.fielden.platform.entity.AbstractPersistentEntity;
-import ua.com.fielden.platform.entity.annotation.KeyType;
-import ua.com.fielden.platform.entity.annotation.KeyTitle;
-import ua.com.fielden.platform.entity.annotation.CompanionObject;
-import ua.com.fielden.platform.entity.annotation.MapEntityTo;
-import ua.com.fielden.platform.entity.annotation.DescTitle;
-import ua.com.fielden.platform.entity.annotation.DisplayDescription;
-import ua.com.fielden.platform.entity.annotation.DescRequired;
-
-/** 
- * One-2-One entity object.
- * 
- * @author Developers
- *
- */
-@KeyType(TgCloseLeaveExample.class)
-@KeyTitle("Key")
-@CompanionObject(ITgCloseLeaveExampleDetail.class)
-@MapEntityTo
-@DescTitle("Desc")
-@DisplayDescription
-@DescRequired
-public class TgCloseLeaveExampleDetail extends AbstractPersistentEntity<TgCloseLeaveExample> {
-
-}
