+++ platform-pojo-bl/src/main/java/fielden/test_app/close_leave/ITgCloseLeaveExample.java
@@ -1,0 +1,13 @@
+package fielden.test_app.close_leave;
+
+import ua.com.fielden.platform.dao.IEntityDao;
+
+/** 
+ * Companion object for entity {@link TgCloseLeaveExample}.
+ * 
+ * @author Developers
+ *
+ */
+public interface ITgCloseLeaveExample extends IEntityDao<TgCloseLeaveExample> {
+
+}
