+++ platform-pojo-bl/src/main/java/fielden/test_app/close_leave/ITgCloseLeaveExampleDetail.java
@@ -1,0 +1,13 @@
+package fielden.test_app.close_leave;
+
+import ua.com.fielden.platform.dao.IEntityDao;
+
+/** 
+ * Companion object for entity {@link TgCloseLeaveExampleDetail}.
+ * 
+ * @author Developers
+ *
+ */
+public interface ITgCloseLeaveExampleDetail extends IEntityDao<TgCloseLeaveExampleDetail> {
+
+}
