+++ platform-pojo-bl/src/main/java/fielden/test_app/close_leave/ITgCloseLeaveExampleMaster_OpenDetailUnpersisted_MenuItem.java
@@ -1,0 +1,13 @@
+package fielden.test_app.close_leave;
+
+import ua.com.fielden.platform.dao.IEntityDao;
+
+/** 
+ * Companion object for entity {@link TgCloseLeaveExampleMaster_OpenDetailUnpersisted_MenuItem}.
+ * 
+ * @author Developers
+ *
+ */
+public interface ITgCloseLeaveExampleMaster_OpenDetailUnpersisted_MenuItem extends IEntityDao<TgCloseLeaveExampleMaster_OpenDetailUnpersisted_MenuItem> {
+
+}
