+++ platform-pojo-bl/src/main/java/fielden/test_app/close_leave/OpenTgCloseLeaveExampleMasterAction.java
@@ -1,0 +1,18 @@
+package fielden.test_app.close_leave;
+
+import ua.com.fielden.platform.entity.AbstractFunctionalEntityToOpenCompoundMaster;
+import ua.com.fielden.platform.entity.annotation.CompanionObject;
+import ua.com.fielden.platform.entity.annotation.KeyTitle;
+import ua.com.fielden.platform.entity.annotation.KeyType;
+/** 
+ * Master entity object.
+ * 
+ * @author Developers
+ *
+ */
+@KeyType(TgCloseLeaveExample.class)
+@KeyTitle("Key")
+@CompanionObject(IOpenTgCloseLeaveExampleMasterAction.class)
+public class OpenTgCloseLeaveExampleMasterAction extends AbstractFunctionalEntityToOpenCompoundMaster<TgCloseLeaveExample> {
+
+}
