+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/web/test/config/ApplicationDomain.java
@@ -6,13 +6,13 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import fielden.close_leave.OpenTgCloseLeaveExampleMasterAction;
-import fielden.close_leave.TgCloseLeaveExample;
-import fielden.close_leave.TgCloseLeaveExampleDetail;
-import fielden.close_leave.TgCloseLeaveExampleDetailUnpersisted;
-import fielden.close_leave.TgCloseLeaveExampleMaster_OpenDetailUnpersisted_MenuItem;
-import fielden.close_leave.TgCloseLeaveExampleMaster_OpenDetail_MenuItem;
-import fielden.close_leave.TgCloseLeaveExampleMaster_OpenMain_MenuItem;
+import fielden.test_app.close_leave.OpenTgCloseLeaveExampleMasterAction;
+import fielden.test_app.close_leave.TgCloseLeaveExample;
+import fielden.test_app.close_leave.TgCloseLeaveExampleDetail;
+import fielden.test_app.close_leave.TgCloseLeaveExampleDetailUnpersisted;
+import fielden.test_app.close_leave.TgCloseLeaveExampleMaster_OpenDetailUnpersisted_MenuItem;
+import fielden.test_app.close_leave.TgCloseLeaveExampleMaster_OpenDetail_MenuItem;
+import fielden.test_app.close_leave.TgCloseLeaveExampleMaster_OpenMain_MenuItem;
 import ua.com.fielden.platform.basic.config.IApplicationDomainProvider;
 import ua.com.fielden.platform.domain.PlatformDomainTypes;
 import ua.com.fielden.platform.entity.AbstractEntity;
