--- platform-web-ui/src/main/java/fielden/main/menu/close_leave/MiTgCloseLeaveExample.java
@@ -1,16 +1,0 @@
-package fielden.main.menu.close_leave;
-
-import fielden.close_leave.TgCloseLeaveExample;
-import ua.com.fielden.platform.entity.annotation.EntityType;
-import ua.com.fielden.platform.ui.menu.MiWithConfigurationSupport;
-
-/**
- * Main menu item representing an entity centre for {@link TgCloseLeaveExample}.
- *
- * @author TG Team
- *
- */
-@EntityType(TgCloseLeaveExample.class)
-public class MiTgCloseLeaveExample extends MiWithConfigurationSupport<TgCloseLeaveExample> {
-
-}
